kernel: Store Multiboot header in first 8 KiB of binary
For the bootloader to recognize the kernel binary, the Multiboot
header must be located within the first 8 KiB of the kernel's ELF
file. The kernel won't be executed otherwise.

This patch fixes the problem by moving the Multiboot header to the
beginning of the text section, and the text section is located first
in the ELF file. The kernel should now be loaded more reliably.