kernel: Introduce list heads
List heads are a special instance of the list data structure. They
represent a list and serve as anchor for the list's entries without
storing an element. It's now possible to easily iterate over the
entries of a list, or insert an entry at the list's end with O(1)
complexity.

This patch changes the list implementation in a fundamental way and
therefore makes invasive changes to the IPC, locking, and scheduling
code.