kernel/i386: Add cli_if_on(), sti_if_on() for conditionally dis-/en-abling IRQs
IRQ may only be disabled once and re-enabled once. Nested calls to change
the IRQ status must succeed without changing the status.

This patch adds cli_if_on() and sti_if_on() that track the IRQ status. A
call to cli_if_on() returns the old IRQ status, which shall be passed to
the corresponding call to sti_if_on(). Both functions do nothing if the
old and the new IRQ status are the same.