kernel/i386: Remove print statements from Multiboot code
It's not save to call console functions from the Multiboot code. The
CRT driver has not yet initialized the device and the console is not
ready.

This patch removes print statements from the Multiboot code. If they
will be required in the future, a simple Multiboot-dependend CRT driver
should be used instead.