kernel: Add pmem_claim_frames() for unconditional frame allocation
The function pmem_claim_frames() unconditionally allocates a range
of page frames starting at a specific location. It does not search
for free frames or check the frame flags for availability.

This functionality is used during the early setup of the physical
memory. Kernel, modules and platform-data is located in the main
memory. To not have this data allocated and overwitten by other
modules, the init code claims all pages that are relevant for the
boot up. All callers have been adapted.