kernel/i386: Claim individual ELF sections of kernel binary
With this patch, the Multiboot code claims individual sections of the
kernel's ELF binary file, instead of claiming the kernel as one large
blob.

When allocating memory, the ELF sections are tested against, instead
of the whole binary. This strategy allows for better memory usage for
kernels with large holes between sections.