Reorganize layout of virtual memory
This patch reorganizes virtual memory. The kernel's memory is now
completely located in the first gigabyte of the address space. There
is a 1024-page area right before 1 GiB for temporary mappings, such
as frames containing page-tables. The user-space memory starts at 1
GiB and ends at the end of the address space. The first page is never
mapped.