kernel/i386: Create kernel-task virtual address space in Multiboot code
The Multiboot code now creates the virtual address space for the kernel
task. The initial layout of this address space contains mappings of
Multiboot data structures, so it makes sense to let Multiboot build the
mapping. Currently, we map the initial 4 MiB unconditionally, but later
patches can implement a more fine-grained mapping.