kernel/i386; Init VMEM in Multiboot code
This patch moves the initialization of the VMEM module to the
Multiboot code. The virtual address space depends on the memory
information that are passed from the Multiboot boot loader.
With the VMEM initialization being located in the Multiboot
code, the initial address space can be created with less memory
overhead.